pyproject.toml
[project]
name = "etsy-assistant"
version = "0.1.0"
description = "Etsy shop assistant for pen & ink sketch artists"
requires-python = ">=3.12"
dependencies = [
    "opencv-python-headless>=4.9",
    "Pillow>=10.0",
    "numpy>=1.26",
    "click>=8.1",
    "anthropic>=0.40",
    "httpx>=0.27",
]

[project.scripts]
etsy-assistant = "etsy_assistant.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/etsy_assistant"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = ["pytest>=8.0"]
